[data, trainer] refactor: merge the multimodal chat templates into one registry
veomni/data/multimodal/multimodal_chat_template.py held a second, hand-written
TEMPLATES dict alongside the CHAT_TEMPLATE_REGISTRY in the parent module, so the
same concept had two registration mechanisms and two build entrypoints that
differed only in whether they forwarded kwargs. With the V1 templates gone the
file was down to three entries, which is not worth a module of its own.

The three templates move verbatim into veomni/data/chat_template.py and register
through the decorator, and build_multimodal_chat_template is dropped in favour
of build_chat_template.

Folding two namespaces into one means a config naming the wrong kind of template
now resolves instead of raising, and would only fail later inside a dataloader
worker on the differing encode_messages signature. build_chat_template therefore
takes an optional expect_multimodal, which the three trainers pass so the
mismatch is reported at build time with the valid alternatives listed.
