+++ docs/usage/basic_modules.md
@@ -246,11 +246,11 @@ VeOmni offers unified multimodal transform functions in [veomni/data/data_transf
 
 Example usage in `_build_data_transform` in [veomni/trainer/vlm_trainer.py](https://github.com/ByteDance-Seed/VeOmni/blob/main/veomni/trainer/vlm_trainer.py).
 ```python
-from veomni.data import build_data_transform, build_multimodal_chat_template
+from veomni.data import build_chat_template, build_data_transform
 from veomni.models import build_processor
 
 processor = build_processor(args.model.tokenizer_path)
-chat_template = build_multimodal_chat_template(args.data.chat_template, processor.tokenizer)
+chat_template = build_chat_template(args.data.chat_template, processor.tokenizer)
 position_id_func = model.get_position_id_func()
 transform = build_data_transform(
     model.config.model_type,
@@ -271,8 +271,8 @@ Multimodal dataset transform follows the similar pipeline:
 
 
 ### Chat Template
-VeOmni default supports several chat template(source code: [veomni/data/chat_template.py](https://github.com/ByteDance-Seed/VeOmni/blob/main/veomni/data/chat_template.py) for text-only model and [veomni/data/multimodal/multimodal_chat_template.py](https://github.com/ByteDance-Seed/VeOmni/blob/main/veomni/data/multimodal/multimodal_chat_template.py) for multimodal model):
-you can add your custom chat template by implementing the `ChatTemplate` class.
+VeOmni default supports several chat templates, text-only and multimodal alike, all registered in [veomni/data/chat_template.py](https://github.com/ByteDance-Seed/VeOmni/blob/main/veomni/data/chat_template.py) and built by name through the single `build_chat_template` entrypoint.
+You can add your custom chat template by implementing the `ChatTemplate` class — or `MultimodalChatTemplate` if it needs the per-modality token counts.
 **Custom Template Implementation**:  
 ```python
 from veomni.data.chat_template import ChatTemplate
