+++ veomni/trainer/text_trainer.py
@@ -77,7 +77,9 @@ class TextTrainer:
             self.base.model_assets = [model_config, self.base.tokenizer]
             self.base.chat_template = None
         else:
-            self.base.chat_template = build_chat_template(args.data.chat_template, self.base.tokenizer)
+            self.base.chat_template = build_chat_template(
+                args.data.chat_template, self.base.tokenizer, expect_multimodal=False
+            )
             self.base.model_assets = [model_config, self.base.chat_template]
 
     def _build_data_transform(self):
